Add some noSideEffect pragmas
For some reason, some stringification procs the only call out to various 'rope' implementations appear to compile with the {.noSideEffect.}  pragma applied to them, while others don't. Furthermore, the ones that compile are almost always much more complex than the
ones that don't, implying that maybe the compiler's analysis is broken. The standard library implementations of 'rope' are not marked as {.noSideEffect.} (which makes sense, since they must refer to the global state in order to determine if caching is enabled or
not).